requirements.txt
flask-sqlalchemy
Flask-Migrate
Flask-WTF
Flask-Login
flask-bcrypt
Faker